+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Adds `Overstrike.isBoolean`.
 - Adds `Overstrike.isNumber`.
 - Adds `Overstrike.isString`.
+- Adds `Overstrike.isFunction`.
 
 ## 0.2.0 (Aug 17, 2014)
 - Adds `Overstrike.new`.